requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
djangorestframework==3.16.1
FreeSWITCH-ESL-Python==1.2
idna==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
